small improvements

+++ HsBot.Logic/CommandModules/Admin.cs
@@ -269,18 +269,35 @@
         [Command("timezone-set")]
         [Summary("timezone-set <identifier>|set your own timezone. Get a list of identifiers with the `{cmdPrefix}timezone-list` command")]
         [RequireUserPermission(GuildPermission.ChangeNickname)]
-        public async Task TimezoneSet(int identifier)
+        public async Task TimezoneSet(int identifier, string otherUser = null)
         {
+            if (otherUser != null && !CurrentUser.GuildPermissions.Administrator)
+            {
+                await Context.Channel.BotResponse("Only Administrators can set the timezone for other users.", ResponseType.error);
+                return;
+            }
+
             await CleanupService.DeleteCommand(Context.Message);
-            var timezones = TimeZoneInfo.GetSystemTimeZones();
-            if (identifier < 1 || identifier > timezones.Count)
+            var timeZones = TimeZoneInfo.GetSystemTimeZones();
+            if (identifier < 1 || identifier > timeZones.Count)
             {
                 await Context.Channel.BotResponse("Wrong timezone index", ResponseType.error);
                 return;
             }
 
-            var tz = timezones[identifier - 1];
-            await TimeZoneLogic.SetTimeZone(Context.Guild, Context.Channel, CurrentUser, tz);
+            var user = CurrentUser;
+            if (otherUser != null)
+            {
+                user = Context.Guild.FindUser(CurrentUser, otherUser);
+                if (user == null)
+                {
+                    await Context.Channel.BotResponse("Unknown user: " + otherUser, ResponseType.error);
+                    return;
+                }
+            }
+
+            var tz = timeZones[identifier - 1];
+            await TimeZoneLogic.SetTimeZone(Context.Guild, Context.Channel, user, tz);
         }
 
         [Command("set-rs-queue-role")]
